refactor(architecture): move currency id schemas to domain packages
CryptoCurrencyIdSchema, TokenCurrencyIdSchema and FiatCurrencyIdSchema
(and their inferred types) move from @shared/schema-primitives to the
packages that own them. A primitives package has no business knowing
about crypto, tokens or fiat.

The crypto and token barrels used to re-export these symbols from
primitives, which made them proxies: two import paths for the same
symbol, and no obvious original provider.

Consumers already importing from @domain/entity-currency-* are
unaffected, since the symbols genuinely moved there. Only five files
referenced them through primitives, four of which were inside the
crypto and token packages themselves.
