+++ domain/entity/contact/package.json
@@ -19,6 +19,8 @@
     "unimported": "pnpm knip --directory ../../.. -W domain/entity/contact"
   },
   "dependencies": {
+    "@domain/entity-currency-crypto": "workspace:*",
+    "@domain/entity-currency-token": "workspace:*",
     "@reduxjs/toolkit": "catalog:",
     "@shared/schema-primitives": "workspace:*",
     "zod": "catalog:"
