+++ domain/entity/currency-crypto/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@domain/entity-currency-unit": "workspace:*",
-    "@shared/schema-primitives": "workspace:*",
     "zod": "catalog:"
   },
   "devDependencies": {
