+++ domain/entity/currency-crypto/src/index.ts
@@ -1,5 +1,3 @@
 export * from "./schema";
 export * from "./registry";
 export * from "./legacy/accessors";
-export type { CryptoCurrencyId } from "@shared/schema-primitives";
-export { CryptoCurrencyIdSchema } from "@shared/schema-primitives";
