+++ domain/entity/currency-token/src/index.ts
@@ -1,4 +1,2 @@
 export * from "./schema";
 export * from "./define";
-export type { TokenCurrencyId } from "@shared/schema-primitives";
-export { TokenCurrencyIdSchema } from "@shared/schema-primitives";
